+++ Slot_Common/common_mbox.h
@@ -39,6 +39,6 @@ typedef struct commonAnalogdata
 }commonAnalogdata_t;
 
 extern commonAnalogdata_t gMboxADMes;
-extern uint16_t gMBoxTempCell;
+extern int32_t gMBoxTempCell;
 
 #endif /* COMMON_MBOX_H_ */
